+++ modules/settings.mjs
@@ -1,4 +1,17 @@
 export default class ModuleSettings {
+  static add(key, data = {}) {
+    const defaults = {
+      scope: 'world',
+      config: true
+    }
+
+    game.settings.register('fvtt-party-resources', key, Object.assign(defaults, data))
+  }
+
+  static get(key) {
+    return game.settings.get('fvtt-party-resources', key)
+  }
+
   static register() {
     window.pr.api.register('resource_list')